Add External Stylesheets
ISSUE: http://redmine-3.2.2.dev/issues/11

The <link> tag in HTML is used to include external stylesheets in a
webpage. In this lesson, we add three external stylesheets.

Normalize.css is a new approach to the CSS
reset. Instead of clearing out all base styles, it attempts to normalize
the look of elements across browsers.

Open Sans is a web font from the Google Fonts library. Web fonts allow
us to include fonts in web pages that are not installed on the browsing
user's desktop.

In this exercise, we include:

* Add normalize.css ()
* Add Open Sans font ()
* Add Ionic Framework icons ()