HTML Markup for Site Navigation and Collection Page
Now that the framework for our home page is complete, we need to create
supplemental site pages. In this checkpoint, we create a new content
page called collection.html and display four album covers on the page.
Next, we modify index.html to include a top navigation container and we
add a link to our Collection page in this container.

* Modify index.html lines 17-22 to include a links container with one link
* Create collection.html with a placeholder for four album covers