Create Landing Page for Albums
Now that we have our collection page, it is time to create a landing
page for an album. When an album cover or title is clicked on our
collection page, it should redirect to album.html. This page should
include an album cover, title, and artist name, release date and
publisher in the header. Below the header, a table should appear that is
populated with album tracks. The table should include track number, name
and duration.

* Create album.html page
* Populate page with top navigation
* Add the album class to the body tag
* Add the main element with a section tag
* Add a table below the section tag