Implement forEach Utility Function and Refactor landing.js
In this assignment, I created a new JavaScript file called utilities.js
and created a forEach utility function that accepts an array and a
function as a parameter. The forEach function then runs the callback
function on each item in the array.

Afterwards, I refactored the landing.js file to utilize this function.