Refactor Gulpfile to Prevent JavaScript Concatenation
The issue was not the renaming of variables in the all.js file. Instead,
the issue was that the script that is supposed to run on the landing
page throws an error when that code is included on the collection page.

On lines 19-20 in landing.js, we attempt to get and work with elements
on the page. Because these elements don't exist in the collection page,
an error is thrown and the JavaScript that generates the albums in
collection view does not process.

To correct this:

* Comment out the line that calls concat() in the scripts task
* Remove browserify as a test to see if it is preventing the watch task
  from running after two edits