Add HTML and CSS for Music Player Controls
In this checkpoint, we add a player-bar section to the
album.html page. This container includes three div elements
for main controls, currently playing song, and volume
controls.

To our project workflow, we add scss-lint and refactor all
scss files in order to remove most warning indicators. We
also create _playerbar.scss to hold the styles for the
player-bar container.

* Use Ionic icons for the music player controls
* Use css to create seek bar and volume slider
* Add .scss-lint.yml configuration file