Allow Users to Toggle Play/Pause State from the Player Bar
In this commit, we create a variable called $playPauseToggle - a
jQuery selector that targets the play/pause toggle button on the
player bar. We use this variable to create a click event handler
in the document.ready function.

When the song is playing, togglePlayFromPlayerBar:

* Sets the play/pause button to a Pause Button state
* Shows a pause button on the song row in the track list
* Pauses the song on click

When the song is paused, togglePlayFromPlayerBar:

* Sets the play/pause button to a Play button state
* Shows the track number in place of the pause button in the
  track list
* Plays the song on click