Restructure BlocJams Project Using AngularJS Framework
Create an app directory to host the core deployment folder
structure for an AngularJS single page application. In this
revised version of BlocJams, the home page, collection page
and album pages will be refactored into a SPA model using
templates.

Here we create the app directory structure:

* assets
* assets/images
* scripts
* styles
* templates
* index.html

We add the Angular.js library via script tag to the HEAD
section of the document, and test our build process.

In this project, I have opted to use Gulp instead of Grunt.