Create PlayerBar Partial and Delete Legacy Assets Directory
Create a partial file that contains the player bar controls and
include it in the album page using ng-include.

Delete the assets directory from the root of the project. This was
the legacy location for the bloc-jams project and is no longer
needed.

Refactor album.html to include AngularJS directives ng-click and
ng-show for interactivity to replace behaviors created using
jQuery in the previous project.