Implement Landing Page and Collection Page Controllers
Create an initial JS controller (LandingCtrl.js) for the
landing page that contains a variable (heroTitle) that contains
the text "Turn the Music Up!". In the landing.html template,
display this variable inside the h3 element that includes the
class hero-title.

Create a fixtures.js file that contains three album objects
which include high-level metadata (title, artist, etc.) and the
list of songs on the album (title, duration, url). In the
CollectionCtrl.js file, create a global variable (albums) and
set it equal to an empty array. Push twelve copies of the first
album from the fixtures file (albumPicasso) into the array.

In the collection.html file, inject the album cover image,
album title, artist and number of tracks using Angular
expressions.

Finally, in app.js, add the controller key to the state method
call on the stateProvider for the landing page and the collection
page.

Modify the gulp build process to automatically generate
JavaScript files in subdirectories so controllers are placed in
a scripts/controllers subfolder.